wip: Add command_name to hook system for plugin command execution
This is a work-in-progress implementation that adds command_name
to PreCommand/PostCommand hooks to allow plugins to intercept
command execution. However, this approach may not be the best
solution - function callbacks would be more direct and ergonomic.

Changes:
- Added command_name field to PreCommand and PostCommand hooks
- Updated plugin_manager to serialize command_name to Lua
- Modified git-grep and git-find-file plugins to use post-command hooks
- Changed plugin commands to use action = "none"

Note: This approach works but is indirect. Better alternatives:
1. Function callbacks in register_command (more direct)
2. String-based function lookup (simpler implementation)
3. Helper wrapper for cleaner API

Next step: Consider refactoring to use function callbacks instead
of hook-based command interception for better plugin ergonomics.